Cleanup slug interface
