+++ app/core/interfaces/slug/input.html
@@ -1,0 +1,3 @@
+<input type="text" value="{{value}}" name="{{name}}" id="{{name}}" maxLength="{{maxLength}}" class="{{size}}" {{#if readonly}}readonly{{/if}}/>
+<span class="char-count hide">{{characters}}</span>
+
